decouple KeyPadHandler from TraditionalT9 - 1
